switch to double precision to avoid to the odd float precision. also stabilize even more temp and humidity. introduce filedatabase to store data hourly and simply to access due http

+++ src/Ens160Aht2x.h
@@ -2,14 +2,14 @@
 
 void Ens160Aht2x_setup();
 void Ens160Aht2x_loop();
-void Ens160Aht2x_setDataListner(void func(float temp, float humidity, int aqi, int tvoc, int eco2));
-float Ens160Aht2x_getTemperature();
-float Ens160Aht2x_getHumidity();
-void Ens160Aht2x_setTempHumDif(float tempdif, float humdif);
-float Ens160Aht2x_getTemperatureDif();
-float Ens160Aht2x_getHumidityDif();
-float Ens160Aht2x_getAvarageTemperature();
-float Ens160Aht2x_getAvarageHumidity();
-float Ens160Aht2x_getVpdAir();
-float Ens160Aht2x_getVpdLeaf();
+void Ens160Aht2x_setDataListner(void func(double temp, double humidity, int aqi, int tvoc, int eco2));
+double Ens160Aht2x_getTemperature();
+double Ens160Aht2x_getHumidity();
+void Ens160Aht2x_setTempHumDif(double tempdif, double humdif);
+double Ens160Aht2x_getTemperatureDif();
+double Ens160Aht2x_getHumidityDif();
+double Ens160Aht2x_getAvarageTemperature();
+double Ens160Aht2x_getAvarageHumidity();
+double Ens160Aht2x_getVpdAir();
+double Ens160Aht2x_getVpdLeaf();
 int Ens160Aht2x_getCo2();
