+++ src/FileController.h
@@ -1,4 +1,4 @@
 #pragma once
 
 void FileController_setup();
-void FileController_write(float temp, float hum, int fanspeed, int co2, int lightmv, float vpd);
+void FileController_write(double temp, double hum, int fanspeed, int co2, int lightmv, double vpd);
